+++ src/utils/index.js
@@ -89,8 +89,6 @@ export function postParser(post) {
     ...post,
     commentsDisabled: post.commentsDisabled === '1',
     savePostStatus: initialAsyncState,
-    // After what comment the omittedComments span is started?
-    omittedCommentsOffset: post.omittedComments > 0 ? 1 : 0,
     // All hashtags used in the post body
     hashtags: tokenizeHashtags()(post.body || '').map((t) => t.text),
   };